+++ src/class.rs
@@ -1,9 +1,23 @@
 //! Contains the class implementation.
 
-use crate::data::usb::constants::*;
-use crate::data::usb_midi::usb_midi_event_packet::{MidiPacketParsingError, UsbMidiEventPacket};
 use usb_device::class_prelude::*;
 use usb_device::Result;
+
+use crate::packet::{UsbMidiEventPacket, UsbMidiEventPacketError};
+
+// Constants for use in descriptors.
+const USB_AUDIO_CLASS: u8 = 0x01;
+const USB_AUDIOCONTROL_SUBCLASS: u8 = 0x01;
+const USB_MIDISTREAMING_SUBCLASS: u8 = 0x03;
+const MIDI_IN_JACK_SUBTYPE: u8 = 0x02;
+const MIDI_OUT_JACK_SUBTYPE: u8 = 0x03;
+const EMBEDDED: u8 = 0x01;
+const EXTERNAL: u8 = 0x02;
+const CS_INTERFACE: u8 = 0x24;
+const CS_ENDPOINT: u8 = 0x25;
+const HEADER_SUBTYPE: u8 = 0x01;
+const MS_HEADER_SUBTYPE: u8 = 0x01;
+const MS_GENERAL: u8 = 0x01;
 
 const MIDI_IN_SIZE: u8 = 0x06;
 const MIDI_OUT_SIZE: u8 = 0x09;
@@ -15,7 +29,7 @@ pub const MIDI_PACKET_SIZE: usize = 4;
 pub const MAX_PACKET_SIZE: usize = 64;
 
 /// Packet-level implementation of a USB MIDI device.
-pub struct MidiClass<'a, B: UsbBus> {
+pub struct UsbMidiClass<'a, B: UsbBus> {
     standard_ac: InterfaceNumber,
     standard_mc: InterfaceNumber,
     standard_bulkout: EndpointOut<'a, B>,
@@ -25,30 +39,31 @@ pub struct MidiClass<'a, B: UsbBus> {
 }
 
 /// Error variants for read operations.
-pub enum MidiReadError {
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum UsbMidiReadError {
     /// Parsing of the packet failed.
-    ParsingFailed(MidiPacketParsingError),
+    ParsingFailed(UsbMidiEventPacketError),
     /// USB stack error returned from `usb-device`.
     UsbError(UsbError),
 }
 
-/// Error returned when passing invalid arguments to `MidiClass::new`.
+/// Error returned when passing invalid arguments to `UsbMidiClass::new`.
 #[derive(Debug)]
 pub struct InvalidArguments;
 
-impl<B: UsbBus> MidiClass<'_, B> {
-    /// Creates a new MidiClass with the provided UsbBus and `n_in/out_jacks` embedded input/output jacks
+impl<B: UsbBus> UsbMidiClass<'_, B> {
+    /// Creates a new UsbMidiClass with the provided UsbBus and `n_in/out_jacks` embedded input/output jacks
     /// (or "cables", depending on the terminology).
     /// Note that a maximum of 16 in and 16 out jacks is supported.
     pub fn new(
         alloc: &UsbBusAllocator<B>,
         n_in_jacks: u8,
         n_out_jacks: u8,
-    ) -> core::result::Result<MidiClass<'_, B>, InvalidArguments> {
+    ) -> core::result::Result<UsbMidiClass<'_, B>, InvalidArguments> {
         if n_in_jacks > 16 || n_out_jacks > 16 {
             return Err(InvalidArguments);
         }
-        Ok(MidiClass {
+        Ok(UsbMidiClass {
             standard_ac: alloc.interface(),
             standard_mc: alloc.interface(),
             standard_bulkout: alloc.bulk(MAX_PACKET_SIZE as u16),
@@ -64,7 +79,7 @@ impl<B: UsbBus> MidiClass<'_, B> {
     }
 
     /// Sends a `UsbMidiEventPacket` and returns either the transferred size or an error.
-    pub fn send_message(&mut self, usb_midi: UsbMidiEventPacket) -> Result<usize> {
+    pub fn send_packet(&mut self, usb_midi: UsbMidiEventPacket) -> Result<usize> {
         let bytes: [u8; MIDI_PACKET_SIZE] = usb_midi.into();
         self.standard_bulkin.write(&bytes)
     }
@@ -99,7 +114,7 @@ impl<B: UsbBus> MidiClass<'_, B> {
     }
 }
 
-impl<B: UsbBus> UsbClass<B> for MidiClass<'_, B> {
+impl<B: UsbBus> UsbClass<B> for UsbMidiClass<'_, B> {
     fn get_configuration_descriptors(&self, writer: &mut DescriptorWriter) -> Result<()> {
         // AUDIO CONTROL STANDARD
         writer.interface(