+++ src/data/midi/mod.rs
@@ -1,5 +1,0 @@
-//! MIDI message and related types.
-
-pub mod channel;
-pub mod message;
-pub mod notes;