+++ src/data/mod.rs
@@ -1,6 +1,0 @@
-//! Primitives and structures used in USB MIDI data.
-
-pub mod byte;
-pub mod midi;
-pub mod usb;
-pub mod usb_midi;