+++ src/data/usb_midi/mod.rs
@@ -1,6 +1,0 @@
-//! USB MIDI event packet parser and related types.
-
-pub mod cable_number;
-pub mod code_index_number;
-pub mod midi_packet_reader;
-pub mod usb_midi_event_packet;