+++ src/message/control_function.rs
@@ -1,9 +1,9 @@
 //! Control function definitions.
 
-use crate::data::byte::u7::U7;
+use crate::message::data::u7::U7;
 
 /// Custom type for a control function.
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub struct ControlFunction(pub U7);
 
 /// Control Functions as defined in the MIDI 1.0 Specification.