+++ src/message/data/u4.rs
@@ -1,12 +1,12 @@
 //! A primitive value with 4-bit length.
 
-use core::convert::TryFrom;
+use crate::message::data::{FromClamped, FromOverFlow};
 
 /// A primitive value that can be from 0-0x0F
 pub struct U4(u8);
 
 /// Error representing that this value is not a valid u4
-pub struct InvalidU4(u8);
+pub struct InvalidU4(pub u8);
 
 impl TryFrom<u8> for U4 {
     type Error = InvalidU4;
@@ -23,6 +23,23 @@ impl TryFrom<u8> for U4 {
 impl From<U4> for u8 {
     fn from(value: U4) -> u8 {
         value.0
+    }
+}
+
+impl FromOverFlow<u8> for U4 {
+    fn from_overflow(value: u8) -> U4 {
+        const MASK: u8 = 0b0000_1111;
+        let value = MASK & value;
+        U4(value)
+    }
+}
+
+impl FromClamped<u8> for U4 {
+    fn from_clamped(value: u8) -> U4 {
+        match U4::try_from(value) {
+            Ok(x) => x,
+            _ => U4::MAX,
+        }
     }
 }