+++ src/message/mod.rs
@@ -1,16 +1,20 @@
 //! Enum representing the MIDI message.
 
+pub mod channel;
 pub mod control_function;
+pub mod data;
+pub mod notes;
 pub mod raw;
 
-use crate::data::byte::from_traits::FromClamped;
-use crate::data::byte::u7::U7;
-use crate::data::midi::channel::Channel;
-use crate::data::midi::message::control_function::ControlFunction;
-use crate::data::midi::message::raw::{Payload, Raw};
-use crate::data::midi::notes::Note;
-use crate::data::usb_midi::usb_midi_event_packet::MidiPacketParsingError;
-use core::convert::TryFrom;
+use crate::message::channel::Channel;
+use crate::message::control_function::ControlFunction;
+use crate::message::data::u7::U7;
+use crate::message::data::FromClamped;
+use crate::message::notes::Note;
+use crate::message::raw::{Payload, Raw};
+use crate::packet::cable_number::CableNumber;
+use crate::packet::code_index_number::CodeIndexNumber;
+use crate::packet::{UsbMidiEventPacket, UsbMidiEventPacketError};
 
 type Velocity = U7;
 
@@ -19,7 +23,7 @@ type Velocity = U7;
 /// Note: not current exhaustive and SysEx messages end up
 /// being a confusing case. So are currently note implemented
 /// they are sort-of unbounded
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, Eq, PartialEq, Clone)]
 pub enum Message {
     /// Note On message.
     NoteOff(Channel, Note, Velocity),
@@ -88,11 +92,12 @@ impl From<Message> for Raw {
 }
 
 impl TryFrom<&[u8]> for Message {
-    type Error = MidiPacketParsingError;
+    type Error = UsbMidiEventPacketError;
+
     fn try_from(data: &[u8]) -> Result<Self, Self::Error> {
         let status_byte = match data.first() {
             Some(byte) => byte,
-            None => return Err(MidiPacketParsingError::MissingDataPacket),
+            None => return Err(UsbMidiEventPacketError::MissingDataPacket),
         };
 
         let event_type = status_byte & 0b1111_0000;
@@ -128,27 +133,111 @@ impl TryFrom<&[u8]> for Message {
                 ControlFunction(get_u7_at(data, 1)?),
                 get_u7_at(data, 2)?,
             )),
-            _ => Err(MidiPacketParsingError::InvalidEventType(event_type)),
+            _ => Err(UsbMidiEventPacketError::InvalidEventType(event_type)),
         }
     }
 }
 
-fn get_note(data: &[u8]) -> Result<Note, MidiPacketParsingError> {
-    let note_byte = get_byte_at_position(data, 1)?;
-    match Note::try_from(note_byte) {
-        Ok(note) => Ok(note),
-        Err(_) => Err(MidiPacketParsingError::InvalidNote(note_byte)),
+impl TryFrom<&UsbMidiEventPacket> for Message {
+    type Error = UsbMidiEventPacketError;
+
+    fn try_from(value: &UsbMidiEventPacket) -> Result<Self, Self::Error> {
+        Self::try_from(&value.as_raw_bytes()[1..])
     }
 }
 
-fn get_u7_at(data: &[u8], index: usize) -> Result<U7, MidiPacketParsingError> {
+impl Message {
+    /// Create a packet from the message.
+    pub fn into_packet(self, cable: CableNumber) -> UsbMidiEventPacket {
+        let cin = self.code_index_number() as u8;
+
+        let mut raw = [0; 4];
+        raw[0] = (cable as u8) << 4 | cin;
+        let r = Raw::from(self);
+        raw[1] = r.status;
+
+        match r.payload {
+            Payload::Empty => {
+                raw[2] = 0;
+                raw[3] = 0;
+            }
+            Payload::SingleByte(byte) => {
+                raw[2] = byte.0;
+                raw[3] = 0;
+            }
+            Payload::DoubleByte(byte1, byte2) => {
+                raw[2] = byte1.0;
+                raw[3] = byte2.0;
+            }
+        };
+
+        UsbMidiEventPacket::try_from(raw.as_slice()).unwrap()
+    }
+
+    /// Returns the code index number for a message.
+    pub fn code_index_number(&self) -> CodeIndexNumber {
+        match self {
+            Self::NoteOn(_, _, _) => CodeIndexNumber::NoteOn,
+            Self::NoteOff(_, _, _) => CodeIndexNumber::NoteOff,
+            Self::ChannelAftertouch(_, _) => CodeIndexNumber::ChannelPressure,
+            Self::PitchWheelChange(_, _, _) => CodeIndexNumber::PitchBendChange,
+            Self::PolyphonicAftertouch(_, _, _) => CodeIndexNumber::PolyKeyPress,
+            Self::ProgramChange(_, _) => CodeIndexNumber::ProgramChange,
+            Self::ControlChange(_, _, _) => CodeIndexNumber::ControlChange,
+        }
+    }
+}
+
+fn get_note(data: &[u8]) -> Result<Note, UsbMidiEventPacketError> {
+    let note_byte = get_byte_at_position(data, 1)?;
+    match Note::try_from(note_byte) {
+        Ok(note) => Ok(note),
+        Err(_) => Err(UsbMidiEventPacketError::InvalidNote(note_byte)),
+    }
+}
+
+fn get_u7_at(data: &[u8], index: usize) -> Result<U7, UsbMidiEventPacketError> {
     let data_byte = get_byte_at_position(data, index)?;
     Ok(U7::from_clamped(data_byte))
 }
 
-fn get_byte_at_position(data: &[u8], index: usize) -> Result<u8, MidiPacketParsingError> {
+fn get_byte_at_position(data: &[u8], index: usize) -> Result<u8, UsbMidiEventPacketError> {
     match data.get(index) {
         Some(byte) => Ok(*byte),
-        None => Err(MidiPacketParsingError::MissingDataPacket),
+        None => Err(UsbMidiEventPacketError::MissingDataPacket),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::message::channel::Channel::{Channel1, Channel2};
+    use crate::message::control_function::ControlFunction;
+    use crate::message::data::u7::U7;
+    use crate::message::notes::Note;
+    use crate::message::Message;
+    use crate::packet::cable_number::CableNumber::{Cable0, Cable1};
+    use crate::packet::UsbMidiEventPacket;
+
+    macro_rules! decode_message_test {
+        ($($id:ident:$value:expr,)*) => {
+            $(
+                #[test]
+                fn $id() {
+                    let (usb_midi_data_packet, expected) = $value;
+                    let message = UsbMidiEventPacket::try_from(&usb_midi_data_packet[..]).unwrap();
+                    assert_eq!(expected, message);
+                }
+            )*
+        }
+    }
+
+    decode_message_test! {
+        note_on: ([9, 144, 36, 127], Message::NoteOn(Channel1, Note::C2, U7(127)).into_packet(Cable0)),
+        note_off: ([8, 128, 36, 0], Message::NoteOff(Channel1, Note::C2, U7(0)).into_packet(Cable0)),
+        polyphonic_aftertouch: ([10, 160, 36, 64], Message::PolyphonicAftertouch(Channel1, Note::C2, U7(64)).into_packet(Cable0)),
+        program_change: ([28, 192, 127, 0], Message::ProgramChange(Channel1, U7(127)).into_packet(Cable1)),
+        channel_aftertouch: ([13, 208, 127, 0], Message::ChannelAftertouch(Channel1, U7(127)).into_packet(Cable0)),
+        pitch_wheel: ([14, 224, 64, 32], Message::PitchWheelChange(Channel1, U7(64), U7(32)).into_packet(Cable0)),
+        control_change: ([11, 177, 1, 32], Message::ControlChange(Channel2, ControlFunction::MOD_WHEEL_1, U7(32)).into_packet(Cable0)),
     }
 }